OTT-939: Refactoring
